refactor: moved to utils
