+++ src/main/java/dcoms/utils/Database.java
@@ -1,4 +1,4 @@
-package dcoms;
+package dcoms.utils;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
@@ -6,7 +6,6 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 
-import dcoms.utils.Env;
 import io.github.cdimascio.dotenv.Dotenv;
 
 public class Database {
